Redid row reduction, I think it's better now
